renamed 'cli_functions' to 'publish_functions'
